third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFEGaussianBlurElement-dom-stdDeviation-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'stdDeviation' attribute of the SVGFEGaussianBlurElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS blurElement.getAttribute('stdDeviation') is "5"
PASS blurElement.getAttribute('stdDeviation') is "10"
PASS successfullyParsed is true

TEST COMPLETE
